+++ emulation/rocjitsu/lib/rocjitsu/src/rocjitsu/code/CMakeLists.txt
@@ -20,6 +20,7 @@ rj_add_object_library(rocjitsu_code
     dbt/semantic_translator.cpp
     dbt/waitcnt_translator.cpp
     patch/code_object_patcher.cpp
+    patch/kernel_text_layout.cpp
     patch/instruction_builder.cpp
     patch/spill_manager.cpp
 )
