+++ emulation/rocjitsu/lib/rocjitsu/src/rocjitsu/code/dbt/binary_translator.cpp
@@ -18,6 +18,7 @@
 #include "rocjitsu/code/dbt/semantic_translator.h"
 #include "rocjitsu/code/patch/code_object_patcher.h"
 #include "rocjitsu/code/patch/instruction_builder.h"
+#include "rocjitsu/code/patch/kernel_text_layout.h"
 #include "rocjitsu/isa/decoder.h"
 #include "rocjitsu/isa/instruction.h"
 
@@ -35,53 +36,7 @@
 
 namespace rocjitsu {
 
-/// @brief Relocated placement of one source CFG block in one emitted kernel.
-///
-/// @details A source block can be emitted more than once when multiple kernel
-/// entries reach shared code. The first relocation implementation rejects shared
-/// blocks, but this struct is still kernel-local so later duplication can reuse
-/// the same fixup model.
-struct BlockPlacement {
-  BasicBlock *block = nullptr; ///< Source CFG block.
-  uint64_t source_start = 0;   ///< Original .text-relative block start.
-  uint64_t source_end = 0;     ///< Original .text-relative block end.
-  uint64_t target_start = 0;   ///< New .text-relative block start.
-  uint64_t target_end = 0;     ///< New .text-relative block end.
-};
-
-/// @brief Pending direct PC-relative branch fixup in one relocated kernel.
-///
-/// @details Source offsets are resolved through the kernel-local placement map
-/// after all reachable blocks and local cave bytes have final target offsets.
-struct BranchFixup {
-  const Instruction *inst = nullptr; ///< Decoded source branch instruction.
-  uint64_t source_inst_offset = 0;   ///< Original .text offset of the branch.
-  uint64_t source_target_offset = 0; ///< Original .text offset of the branch target.
-  uint64_t target_inst_offset = 0;   ///< New .text offset of the branch instruction.
-};
-
-/// @brief Physical output layout for one translated kernel.
-///
-/// @details Blocks are emitted in original .text order. This is intentional: it
-/// preserves every CFG fallthrough edge as physical adjacency, so DBT only
-/// patches explicit PC-relative branch immediates. Expansion bodies are appended
-/// after the emitted body as a local cave.
-struct KernelTextLayout {
-  KdTranslation *translation = nullptr;   ///< Descriptor plan for this kernel.
-  uint64_t source_entry = 0;              ///< Original descriptor entry offset.
-  uint64_t target_entry = 0;              ///< Final descriptor entry offset.
-  uint64_t target_body_entry = 0;         ///< Relocated original entry offset.
-  uint64_t body_begin = 0;                ///< First emitted body byte.
-  uint64_t body_end = 0;                  ///< One-past-end of emitted body.
-  uint64_t cave_begin = 0;                ///< First local cave byte.
-  uint64_t cave_end = 0;                  ///< One-past-end of local cave.
-  std::vector<BlockPlacement> blocks;     ///< Kernel-local block placements.
-  std::vector<BranchFixup> branch_fixups; ///< Explicit branch patches.
-};
-
 namespace {
-
-inline constexpr uint64_t kKernargPreloadSkipBytes = 256;
 
 EncodingTranslateFn select_encoding_translator(rj_code_arch_t guest, rj_code_arch_t host) {
   if (guest == ROCJITSU_CODE_ARCH_CDNA4 && host == ROCJITSU_CODE_ARCH_RDNA4)
@@ -153,96 +108,6 @@ LegalizationLookupFn select_legalization(rj_code_arch_t guest, rj_code_arch_t ho
   if (!raw)
     return {};
   return {raw, raw + inst.size() / sizeof(uint32_t)};
-}
-
-void append_words(std::vector<uint8_t> &text, std::span<const uint32_t> words) {
-  const auto *bytes = reinterpret_cast<const uint8_t *>(words.data());
-  text.insert(text.end(), bytes, bytes + words.size() * sizeof(uint32_t));
-}
-
-void append_nop_padding(std::vector<uint8_t> &text, uint64_t byte_count, rj_code_arch_t arch) {
-  assert(byte_count % sizeof(uint32_t) == 0 && "padding must be word-aligned");
-  for (uint64_t off = 0; off < byte_count; off += sizeof(uint32_t)) {
-    const uint32_t nop = build_s_nop(0, arch);
-    append_words(text, std::span<const uint32_t>(&nop, 1));
-  }
-}
-
-[[nodiscard]] uint64_t padding_for_residue(uint64_t current_offset, uint64_t target_residue,
-                                           uint64_t alignment) {
-  const uint64_t current_residue = current_offset % alignment;
-  return (target_residue + alignment - current_residue) % alignment;
-}
-
-[[nodiscard]] uint64_t preserve_entry_skip_window_offset(const KernelTextLayout &layout,
-                                                         const BasicBlock &block, uint64_t cursor) {
-  if (block.start_offset() < layout.source_entry)
-    return cursor;
-
-  // Kernels with kernarg preloading have two valid hardware entry paths:
-  // incompatible firmware executes the descriptor entry and branches around the
-  // preload compatibility prologue, while compatible firmware adds 256 bytes to
-  // KERNEL_CODE_ENTRY_BYTE_OFFSET and starts at the real body directly. Local
-  // body compaction must therefore keep the first 256 bytes after the descriptor
-  // entry stable. Otherwise compatible firmware lands in the middle of compacted
-  // translated code with the preloaded SGPR ABI assumed but not honored.
-  const uint64_t source_delta = block.start_offset() - layout.source_entry;
-  if (source_delta > kKernargPreloadSkipBytes)
-    return cursor;
-
-  return std::max(cursor, layout.body_begin + source_delta);
-}
-
-[[nodiscard]] std::optional<uint64_t> target_for_source_offset(const KernelTextLayout &layout,
-                                                               uint64_t source_offset) {
-  for (const BlockPlacement &placement : layout.blocks) {
-    if (source_offset >= placement.source_start && source_offset < placement.source_end)
-      return placement.target_start + (source_offset - placement.source_start);
-  }
-  return std::nullopt;
-}
-
-[[nodiscard]] uint32_t text_word_at(std::span<const uint8_t> text, uint64_t offset) {
-  uint32_t word = 0;
-  if (offset + sizeof(word) <= text.size())
-    std::memcpy(&word, text.data() + offset, sizeof(word));
-  return word;
-}
-
-[[nodiscard]] bool is_nop_padding_gap(std::span<const uint8_t> text, uint64_t begin, uint64_t end,
-                                      rj_code_arch_t arch) {
-  if (begin > end || end > text.size() || (begin % sizeof(uint32_t)) != 0 ||
-      (end % sizeof(uint32_t)) != 0)
-    return false;
-
-  const uint32_t nop = build_s_nop(0, arch);
-  for (uint64_t off = begin; off < end; off += sizeof(uint32_t)) {
-    if (text_word_at(text, off) != nop)
-      return false;
-  }
-  return true;
-}
-
-[[nodiscard]] std::optional<uint64_t> target_for_source_fallthrough(const KernelTextLayout &layout,
-                                                                    std::span<const uint8_t> text,
-                                                                    uint64_t source_offset,
-                                                                    rj_code_arch_t arch) {
-  if (auto target = target_for_source_offset(layout, source_offset))
-    return target;
-
-  // Local cave expansions return to the source instruction's next PC. When the
-  // compiler placed only s_nop padding between two reachable blocks, that next
-  // PC is not present in the compact relocated body. Skipping that source-only
-  // padding here keeps the body layout compact while preserving the observable
-  // fallthrough target: executing zero or more nops reaches the next block.
-  for (const BlockPlacement &placement : layout.blocks) {
-    if (source_offset >= placement.source_start)
-      continue;
-    if (is_nop_padding_gap(text, source_offset, placement.source_start, arch))
-      return placement.target_start;
-    break;
-  }
-  return std::nullopt;
 }
 
 [[nodiscard]] bool words_changed(std::span<const uint32_t> before,
@@ -485,6 +350,8 @@ TranslatedCodeObject BinaryTranslator::translate(const AmdGpuCodeObject &obj) {
       if (block == nullptr)
         continue;
       if (!translated_blocks.insert(block).second) {
+        // Reject shared source blocks to avoid generating an invalid layout when
+        // one block is used by multiple kernel entry points.
         append_error(result.diagnostics, DiagnosticKind::Legalization,
                      "basic block is reachable from multiple kernel entries; shared kernel text "
                      "translation is not implemented");
